third_party/WebKit/LayoutTests/http/tests/inspector/security/interstitial-sidebar.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../security-test.js"></script>
<script>
function test()
{

    var request1 = new SDK.NetworkRequest(InspectorTest.mainTarget, 0, "https://foo.test/", "https://foo.test", 0, 0, null);
    request1.setSecurityState(Protocol.Security.SecurityState.Secure);
    InspectorTest.dispatchRequestFinished(request1);

    var request2 = new SDK.NetworkRequest(InspectorTest.mainTarget, 0, "https://bar.test/foo.jpg", "https://bar.test", 0, 0, null);
    request2.setSecurityState(Protocol.Security.SecurityState.Secure);
    InspectorTest.dispatchRequestFinished(request2);

    InspectorTest.addResult("Before interstitial is shown:");
    InspectorTest.dumpDeepInnerHTML(Security.SecurityPanel._instance()._sidebarTree.element);

    // Test that the sidebar is hidden when an interstitial is shown. https://crbug.com/559150
    InspectorTest.mainTarget.model(SDK.ResourceTreeModel).dispatchEventToListeners(SDK.ResourceTreeModel.Events.InterstitialShown);
    // Simulate a request finishing after the interstitial is shown, to make sure that doesn't show up in the sidebar.
    var request3 = new SDK.NetworkRequest(InspectorTest.mainTarget, 0, "https://bar.test/foo.jpg", "https://bar.test", 0, 0, null);
    request3.setSecurityState(Protocol.Security.SecurityState.Unknown);
    InspectorTest.dispatchRequestFinished(request3);
    InspectorTest.addResult("After interstitial is shown:");
    InspectorTest.dumpDeepInnerHTML(Security.SecurityPanel._instance()._sidebarTree.element);

    // Test that the sidebar is shown again when the interstitial is hidden. https://crbug.com/559150
    InspectorTest.mainTarget.model(SDK.ResourceTreeModel).dispatchEventToListeners(SDK.ResourceTreeModel.Events.InterstitialHidden);
    InspectorTest.addResult("After interstitial is hidden:");
    InspectorTest.dumpDeepInnerHTML(Security.SecurityPanel._instance()._sidebarTree.element);

    InspectorTest.completeTest();
}
</script>
</head>
<body onload="runTest()">
<p>Tests that the sidebar origin list disappears and appers when an interstitial is shown or hidden.</p>
</body>
</html>
