third_party/WebKit/LayoutTests/http/tests/inspector/security/interstitial-sidebar-expected.txt

Tests that the sidebar origin list disappears and appers when an interstitial is shown or hidden.

Before interstitial is shown:
<DIV >
    <#document-fragment >
        <STYLE type=text/css >
        </STYLE>
        <STYLE type=text/css >
        </STYLE>
        <STYLE type=text/css >
        </STYLE>
        <DIV class=tree-outline-disclosure >
            <OL class=tree-outline tabindex=0 >
                <LI class=security-main-view-sidebar-tree-item selected force-white-icons >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
                    </SPAN>
                    <DIV class=icon lock-icon lock-icon-unknown >
                    </DIV>
                    <SPAN class=title >
Overview
                    </SPAN>
                </LI>
                <OL class=children >
                </OL>
                <LI class=parent expanded security-sidebar-origins >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
Main Origin
                    </SPAN>
                </LI>
                <OL class=children expanded >
                    <LI class=security-main-view-reload-message >
                        <DIV class=selection fill >
                        </DIV>
                        <SPAN class=tree-element-title >
Reload to view details
                        </SPAN>
                    </LI>
                    <OL class=children >
                    </OL>
                </OL>
                <LI class=parent expanded security-sidebar-origins hidden >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
Non-Secure Origins
                    </SPAN>
                </LI>
                <OL class=children expanded hidden >
                </OL>
                <LI class=parent expanded security-sidebar-origins >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
Secure Origins
                    </SPAN>
                </LI>
                <OL class=children expanded >
                    <LI class=security-sidebar-tree-item >
                        <DIV class=selection fill >
                        </DIV>
                        <SPAN class=tree-element-title >
                        </SPAN>
                        <DIV class=icon security-property security-property-secure >
                        </DIV>
                        <SPAN class=title >
https://foo.test
                        </SPAN>
                    </LI>
                    <OL class=children >
                    </OL>
                    <LI class=security-sidebar-tree-item >
                        <DIV class=selection fill >
                        </DIV>
                        <SPAN class=tree-element-title >
                        </SPAN>
                        <DIV class=icon security-property security-property-secure >
                        </DIV>
                        <SPAN class=title >
https://bar.test
                        </SPAN>
                    </LI>
                    <OL class=children >
                    </OL>
                </OL>
                <LI class=parent expanded security-sidebar-origins hidden >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
Unknown / Canceled
                    </SPAN>
                </LI>
                <OL class=children expanded hidden >
                </OL>
            </OL>
        </DIV>
        <STYLE type=text/css >
        </STYLE>
        <STYLE type=text/css >
        </STYLE>
    </#document-fragment>
</DIV>
After interstitial is shown:
<DIV >
    <#document-fragment >
        <STYLE type=text/css >
        </STYLE>
        <STYLE type=text/css >
        </STYLE>
        <STYLE type=text/css >
        </STYLE>
        <DIV class=tree-outline-disclosure >
            <OL class=tree-outline tabindex=0 >
                <LI class=security-main-view-sidebar-tree-item selected force-white-icons >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
                    </SPAN>
                    <DIV class=icon lock-icon lock-icon-unknown >
                    </DIV>
                    <SPAN class=title >
Overview
                    </SPAN>
                </LI>
                <OL class=children >
                </OL>
                <LI class=parent expanded security-sidebar-origins hidden >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
Main Origin
                    </SPAN>
                </LI>
                <OL class=children expanded hidden >
                    <LI class=security-main-view-reload-message >
                        <DIV class=selection fill >
                        </DIV>
                        <SPAN class=tree-element-title >
Reload to view details
                        </SPAN>
                    </LI>
                    <OL class=children >
                    </OL>
                </OL>
                <LI class=parent expanded security-sidebar-origins hidden >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
Non-Secure Origins
                    </SPAN>
                </LI>
                <OL class=children expanded hidden >
                </OL>
                <LI class=parent expanded security-sidebar-origins hidden >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
Secure Origins
                    </SPAN>
                </LI>
                <OL class=children expanded hidden >
                    <LI class=security-sidebar-tree-item >
                        <DIV class=selection fill >
                        </DIV>
                        <SPAN class=tree-element-title >
                        </SPAN>
                        <DIV class=icon security-property security-property-secure >
                        </DIV>
                        <SPAN class=title >
https://foo.test
                        </SPAN>
                    </LI>
                    <OL class=children >
                    </OL>
                    <LI class=security-sidebar-tree-item >
                        <DIV class=selection fill >
                        </DIV>
                        <SPAN class=tree-element-title >
                        </SPAN>
                        <DIV class=icon security-property security-property-secure >
                        </DIV>
                        <SPAN class=title >
https://bar.test
                        </SPAN>
                    </LI>
                    <OL class=children >
                    </OL>
                </OL>
                <LI class=parent expanded security-sidebar-origins hidden >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
Unknown / Canceled
                    </SPAN>
                </LI>
                <OL class=children expanded hidden >
                </OL>
            </OL>
        </DIV>
        <STYLE type=text/css >
        </STYLE>
        <STYLE type=text/css >
        </STYLE>
    </#document-fragment>
</DIV>
After interstitial is hidden:
<DIV >
    <#document-fragment >
        <STYLE type=text/css >
        </STYLE>
        <STYLE type=text/css >
        </STYLE>
        <STYLE type=text/css >
        </STYLE>
        <DIV class=tree-outline-disclosure >
            <OL class=tree-outline tabindex=0 >
                <LI class=security-main-view-sidebar-tree-item selected force-white-icons >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
                    </SPAN>
                    <DIV class=icon lock-icon lock-icon-unknown >
                    </DIV>
                    <SPAN class=title >
Overview
                    </SPAN>
                </LI>
                <OL class=children >
                </OL>
                <LI class=parent expanded security-sidebar-origins >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
Main Origin
                    </SPAN>
                </LI>
                <OL class=children expanded >
                    <LI class=security-main-view-reload-message >
                        <DIV class=selection fill >
                        </DIV>
                        <SPAN class=tree-element-title >
Reload to view details
                        </SPAN>
                    </LI>
                    <OL class=children >
                    </OL>
                </OL>
                <LI class=parent expanded security-sidebar-origins >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
Non-Secure Origins
                    </SPAN>
                </LI>
                <OL class=children expanded >
                </OL>
                <LI class=parent expanded security-sidebar-origins >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
Secure Origins
                    </SPAN>
                </LI>
                <OL class=children expanded >
                    <LI class=security-sidebar-tree-item >
                        <DIV class=selection fill >
                        </DIV>
                        <SPAN class=tree-element-title >
                        </SPAN>
                        <DIV class=icon security-property security-property-secure >
                        </DIV>
                        <SPAN class=title >
https://foo.test
                        </SPAN>
                    </LI>
                    <OL class=children >
                    </OL>
                    <LI class=security-sidebar-tree-item >
                        <DIV class=selection fill >
                        </DIV>
                        <SPAN class=tree-element-title >
                        </SPAN>
                        <DIV class=icon security-property security-property-secure >
                        </DIV>
                        <SPAN class=title >
https://bar.test
                        </SPAN>
                    </LI>
                    <OL class=children >
                    </OL>
                </OL>
                <LI class=parent expanded security-sidebar-origins >
                    <DIV class=selection fill >
                    </DIV>
                    <SPAN class=tree-element-title >
Unknown / Canceled
                    </SPAN>
                </LI>
                <OL class=children expanded >
                </OL>
            </OL>
        </DIV>
        <STYLE type=text/css >
        </STYLE>
        <STYLE type=text/css >
        </STYLE>
    </#document-fragment>
</DIV>
